+++ runner/create_transformation_test.go
@@ -39,7 +39,7 @@ func (m MockOfflineCreateTransformationFail) CreateTrainingSet(provider.Training
 func (m MockOfflineCreateTransformationFail) GetTrainingSet(id provider.ResourceID) (provider.TrainingSetIterator, error) {
 	return nil, nil
 }
-func (m MockOfflineCreateTransformationFail) GetTrainingSetTestSplit(id provider.ResourceID, testSize float64, shuffle bool, randomState int) (provider.TrainingSetIterator, provider.TrainingSetIterator, func() error, error) {
+func (m MockOfflineCreateTransformationFail) GetTrainingSetTestSplit(id provider.ResourceID, testSize float32, shuffle bool, randomState int) (provider.TrainingSetIterator, provider.TrainingSetIterator, func() error, error) {
 	return nil, nil, nil, nil
 }
