+++ runner/register_source_test.go
@@ -41,7 +41,7 @@ func (m MockOfflineRegisterSourceFail) GetTrainingSet(id provider.ResourceID) (p
 	return nil, nil
 }
 
-func (m MockOfflineRegisterSourceFail) GetTrainingSetTestSplit(id provider.ResourceID, testSize float64, shuffle bool, randomState int) (provider.TrainingSetIterator, provider.TrainingSetIterator, func() error, error) {
+func (m MockOfflineRegisterSourceFail) GetTrainingSetTestSplit(id provider.ResourceID, testSize float32, shuffle bool, randomState int) (provider.TrainingSetIterator, provider.TrainingSetIterator, func() error, error) {
 	return nil, nil, nil, nil
 }
