+++ runner/training_set_runner_test.go
@@ -41,7 +41,7 @@ func (m MockOfflineCreateTrainingSetFail) GetTrainingSet(id provider.ResourceID)
 	return nil, nil
 }
 
-func (m MockOfflineCreateTrainingSetFail) GetTrainingSetTestSplit(id provider.ResourceID, testSize float64, shuffle bool, randomState int) (provider.TrainingSetIterator, provider.TrainingSetIterator, func() error, error) {
+func (m MockOfflineCreateTrainingSetFail) GetTrainingSetTestSplit(id provider.ResourceID, testSize float32, shuffle bool, randomState int) (provider.TrainingSetIterator, provider.TrainingSetIterator, func() error, error) {
 	return nil, nil, nil, nil
 }
